fix(FocusManager): Replaces implementation with focus-trap-react (#28)
## Because
In order to fix the bug reported on issue 2

## This Commit
Does a re-implementation of the `FocusManager`, by replacing the current functionality, with the functionality provided by the ["focus-trap-react" package](https://www.npmjs.com/package/focus-trap-react).

In order to keep the same API as before, it makes them still available as before, but changes the default values for `autoFocus`, `restoreFocus` and `contain` props. It also exposes the "focus-trap" configuration API on the `options` prop.

As such, the component itself also gains a whole set of new functionalities, made possible by the `focus-trap` package.

BREAKING CHANGE: autoFocus, restoreFocus and contain are now set to true by default
BREAKING CHANGE: the `useFocusManager` was reimplemented. Since we no longer use a React context state management solution to travel between elements, the hook was re-implemented as an optional way to facilitate the creation of a `FocusTrap` instance, but without using the provided element by the package.

Closes #27

Co-authored-by: João Dias <joao.dias@feedzai.com>